Simplify high-level linting logic
The collaboration between `StyleChecker` and `Linter::Collection`
introduced additional complexity.

The new implementation feels easier to understand and is less code.
