+++ app/models/hound_config.rb
@@ -1,5 +1,17 @@
 # frozen_string_literal: true
 class HoundConfig
+  LINTERS = [
+    Linter::CoffeeScript,
+    Linter::Eslint,
+    Linter::Go,
+    Linter::Haml,
+    Linter::Jshint,
+    Linter::Remark,
+    Linter::Python,
+    Linter::Ruby,
+    Linter::Scss,
+    Linter::Swift,
+  ].freeze
   BETA_LINTERS = %w(
     eslint
     remark
@@ -8,6 +20,10 @@ class HoundConfig
   CONFIG_FILE = ".hound.yml"
 
   attr_reader_initialize :commit
+
+  def self.linter_names
+    LINTERS.map { |klass| klass.name.demodulize.underscore }
+  end
 
   def content
     @_content ||= default_config.deep_merge(resolved_conflicts_config)
@@ -27,7 +43,7 @@ class HoundConfig
   private
 
   def default_config
-    Linter::Collection.linter_names.each.with_object({}) do |name, config|
+    self.class.linter_names.each.with_object({}) do |name, config|
       config[name] = { "enabled" => !BETA_LINTERS.include?(name) }
     end
   end
